refactor: organizando pastas por dominio
