+++ src/main/java/com/corcino/bookstore/generic/Auditable.java
@@ -1,4 +1,4 @@
-package com.corcino.bookstore.model;
+package com.corcino.bookstore.generic;
 
 import lombok.Getter;
 import lombok.Setter;
