+++ src/main/java/com/corcino/bookstore/publisher/repository/PublisherRepository.java
@@ -1,6 +1,6 @@
-package com.corcino.bookstore.repository;
+package com.corcino.bookstore.publisher.repository;
 
-import com.corcino.bookstore.model.Publisher;
+import com.corcino.bookstore.publisher.model.Publisher;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
