+++ src/main/java/com/corcino/bookstore/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.corcino.bookstore.repository;
+package com.corcino.bookstore.user.repository;
 
-import com.corcino.bookstore.model.User;
+import com.corcino.bookstore.user.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
